PAYARA-2598 Move CertificateCredential to Payara API
