+++ api/payara-api/pom.xml
@@ -151,6 +151,7 @@
                                     fish.payara.notification.healthcheck,
                                     fish.payara.notification.requesttracing,
                                     fish.payara.security.annotations,
+                                    fish.payara.security.api,
                                     fish.payara.security.oauth2.api,
                                     fish.payara.security.openid.api
                                 </Export-Package>
